Refactor SchemaError
